pyproject.toml
[project]
name = "daisy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anemone-daisy-maker>=1.95",
    "cn2an>=0.5.23",
    "funasr>=1.2.7",
    "inflect>=7.5.0",
    "jieba>=0.42.1",
    "jiwer==3.1.0",
    "k2>=0.3.0",
    "lhotse>=1.30.3",
    "librosa>=0.11.0",
    "lxml>=6.0.1",
    "numpy>=2.2.6",
    "pyannote-audio>=3.3.2",
    "pyaudio>=0.2.14",
    "pydub>=0.25.1",
    "pymupdf>=1.26.4",
    "pypinyin>=0.55.0",
    "s3prl>=0.4.18",
    "safetensors>=0.6.2",
    "setuptools>=80.9.0",
    "soundfile>=0.13.1",
    "tensorboard>=2.20.0",
    "torch==2.8",
    "torchaudio",
    "tqdm>=4.67.1",
    "transformers>=4.56.1",
    "vocos>=0.1.0",
    "zhconv>=1.4.3",
    "zhon>=2.1.1",
    "piper-phonemize @ https://github.com/csukuangfj/piper-phonemize/releases/download/2025.06.23/piper_phonemize-1.3.0-cp312-cp312-win_amd64.whl",
    "ffmpeg>=1.4",
    "fastapi>=0.116.1",
    "uvicorn>=0.35.0",
    "streamlit>=1.49.1",
    "requests>=2.32.5",
    "python-multipart>=0.0.20",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu126" },
]
torchaudio = [
  { index = "pytorch-cu126" },
]



[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true
